modularize source files, rename source files and module names
